add: eip7702auth to getUserOperation() function if available, removing the need to have it at other places
