+++ src/libs/userOperation/types.ts
@@ -40,6 +40,15 @@ export interface UserOperation {
   // which bundler is responsible for submitting and fetching info
   // about this userOp
   bundler: BUNDLER
+  eip7702auth?: {
+    contractAddress: Hex
+    chainId: Hex
+    nonce: Hex
+    r: Hex
+    s: Hex
+    v: Hex
+    yParity: Hex
+  }
 }
 
 export interface UserOperationEventData {
